requirements.txt
aioboto3~=8.0.5
config-client
faust~=1.10.4
asyncpg
aioredis~=1.3.1
tortoise-orm
Pillow~=7.2.0
numpy~=1.19.1
botocore~=1.17.28